rename paket -> plan
